refactor(meeting_backend): extract sanitize helpers (#1266)
mod.rs 743 -> 587 + sanitize.rs 164. Sanitization helpers
(extract_response, sanitize_agent_output, strip_ansi_escapes,
is_*_line/_block_open/_close) moved to sanitize.rs and re-exported.
mod.rs still over 400 — further impl-method extraction needed in follow-up.
